test: cover random patch error states
Agent-Logs-Url: https://github.com/cat2151/clap-mml-render-tui/sessions/0585b820-d5db-4954-a0cd-214b204c5884

+++ app/src/tui/input.rs
@@ -49,8 +49,8 @@ impl<'a> TuiApp<'a> {
                 let ns = SystemTime::now()
                     .duration_since(UNIX_EPOCH)
                     .map(|duration| duration.as_nanos())
-                    .unwrap_or(0) as usize;
-                let index = ns % pairs.len();
+                    .unwrap_or(0);
+                let index = (ns % pairs.len() as u128) as usize;
                 Ok(pairs[index].0.clone())
             }
         }
